Revert "Now registers doc schedule"
This reverts commit ea2d4e7313e74f762ce026b32212c9301622ce81.

+++ TomcatCapstoneProject/AspGroupOneCapstoneTC/src/main/java/org/aspgroup1/crud/ScheduleCrud.java
@@ -13,7 +13,7 @@ public class ScheduleCrud {
     
     public ScheduleCrud(){}
     
-    public void createSchedule(long doctorID, int mon, int tue, int wed, int thur, int fri, int sat, int sun){
+    public void createSchedule(int doctorID, int mon, int tue, int wed, int thur, int fri, int sat, int sun){
         Schedule schObj;
         Session sessionObj = HibernateUtil.getSessionFactory().openSession();
